Rename files to meet styleguide
