+++ src/algorithms/Algorithms.h
@@ -9,5 +9,3 @@
 #include "algorithms/Pyro.h"
 #include "algorithms/TaneX.h"
 
-/* Algorithms for workflow for typo discovery */
-#include "algorithms/TypoMiner.h"
